+++ db/migrate/20210411195956_add_email_to_users.rb
@@ -1,0 +1,5 @@
+class AddEmailToUsers < ActiveRecord::Migration[6.1]
+  def change
+    add_column :users, :email, :string, null: false, default: ''
+  end
+end