+++ app/helpers/editor/users_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# this module contain specific helpers for {Editor::UsersController}
+module Editor::UsersHelper
+end