+++ app/helpers/home_helper.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+# this module contain specific helpers for {HomeController}
 module HomeHelper
 end