+++ test/mailers/previews/ticket_mailer_preview.rb
@@ -1,4 +1,5 @@
+# frozen_string_literal: true
+
 # Preview all emails at http://localhost:3000/rails/mailers/ticket_mailer
 class TicketMailerPreview < ActionMailer::Preview
-
 end