+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0]
+### Added
+- Users' ticket manage for editor
+
 ## [2.0.0]
 ### Added
 - Add tickets frequency: in the fact model you can select how often a ticket is bookable